Defines XSecScreen GUI Class for Geoms with XSec Tab
This defines the class, derived from GeomScreen, but no screens have
are updated to use this new class in this commit. This class will only
be used to create the XSec tab and control the XSec type section of it.
Special considerations are made for WingGeoms, which use "Airfoil"
instead of "XSec". In addition, a flag is available to control when the
new XSec driver parameterization should be made available.