Implements XSecScreen Class for All Applicable XSec GUIs (Not BOR Geom)
This changes Fuselage, Prop, Stack, and Wing Screen to be derived from
XSecScreen since they are virtually identical. Some differences include
loop treatment for Fuselage and Stack, unavailability of Point type XSec
for Props, and a call to deactivate width and/or height Parms when the
EDIT_CURVE GUI is launched. This also includes some fixes, such as the
missing calls to update Four Series parameters in PropScreen. Another
improvement is the input of enum types into AddItem to avoid the need
for an offset index in PropScreen, which skips Point type XSecs. The
driver groups for width, height, area, and height to width ratio
parameters are only made available for Fuselage and Stack.
RebuildCSTGroup and DisplayGroup are moved to the base class.