Forces Width and Height Drivers for Prop and Wing with EDIT_CURVE XSecs
This adds a function that takes the Geom type as input to force the
drives to Width and Height since Wings and Props use other parameters
to control these values (i.e. chord). The sliders are deactivated if
they can't be set from the XSec editor. Note, the user can still use
invalid drivers from the API at this time.