Adds Carlson Cl Max Option for VSPAERO
The ClMaxToggle BoolParm is changed to an IntParm and the associated GUI
toggle button is turned to a Choice dropdown. Note, the Parm name is
unchanged for backwards compatibility. No changes are needed in
AnalysisMgr since BoolParms are registered as ints. The input can now
be set to the CLMAX_CARLSON enum (2). If this option is selected, Cl Max
is set to 999 to signal to VSPAERO to use this stall modeling type.