Fix what appears to be a copy/paste bug in EditCurveXSec
Not sure that anyone has hit this bug, but it seems this is the right
thing to do from inspection.

Only occurs when using an EditCurveXSec in either Line or PChip mode
and Splitting the curve.  Further, the split must have occurred past
the end of the curve -- which does not seem possible.