Add flags to AddMeshGeom and friends
suppressdisks prevents any surface marked as a DISK_SURF from being added
when AddMeshGeom is called.  This is needed by VSPAERO, where the geometry
for an actuator disk is passed by center, normal, diameter instead of
triangles.

hideoffset controls whether the target Set is hidden when the MeshGeom
is created from CompGeom or WriteVSPGeomFile.

Hiding this Geom is causing the VSPAERO GUI to eliminate actuator disks
when analyzing the Shown set.  We turn it off to keep them Shown.