OpenVSP 3.26.1
If 3.26.0 wasn't new enough for you, here is one better.  No new features
here, but a couple of bug fixes that are pretty essential.

If you weren't on 3.26.0, you should because it has lots of good stuff,
If you're on 3.26.0, you should move to 3.26.1 because it has some
important fixes.

Most importantly, a list-moment change to 3.26.0 disabled a few buttons
and other things in the GUI -- Wing Airfoil Copy/Paste, changing which
Prop Curve you were editing, and two new Prop curve features
Approximate All and Reset Thickness.  These fixes alone are worth
the price of admission.

In addition, I've finally (knock on wood) fixed the issues that were
making actuator disk not work from the GUI with thick-surface files,
thin-surface via the experimental format, and general odd behavior
when using the Shown set.

Plus a fix for a bug that might be impossible to hit.

Also, some more spelling corrections from Luz Paz -- for people who care
about spelling errors in the comments of source code.  Is my mother
watching?  Thanks for the contibution.

Features:
 - None

Bug Fixes:
 - Fix broken GUI functions (Wing Airfoil copy/paste & Prop curve stuff)
 - Fix actuator disk broken cases with VSPAERO
 - Other small fixes

Cleanups:
 - Another round of source comment spelling fixes - thanks Luz Paz