Add u-driven RoundJoints to VspCurve.
Sometimes it is easier/better to identify corners by parameter rather than
index.

Although it was initially attractive to re-use the pmap, that became
impractical as the corner rounding an Chevron needs increased in complexity.

In the end, we could not guarantee a stable pmap as corners are rounded,
so it makes sense to just get a fresh pmap each time through.

Also use tolerance version of vector_find_val to allow some imperfection
in inputs.