Correct order of dihedral and twist transformations
Thanks to Jeff Viken for pointing out this error.

Airfoil rotation to match dihedral needs to be applied before twist.
Since transformations are applied in the reverse order of the code,
they need to be reversed here.

The Z direction here is always zero.  It could be omitted.  However, as
it might be applied later, it is left in for now.  Its relative position
is kept the same (XYZ to ZYX) as YZX or YXZ have no justification.