Fix bug where curve parameter did not match surface parameter.
Apparently, CFDMesh never really used the curve parameter interchangeably
with surface parameter.  Instead it operated on control points directly,
or just used the curve parameter in a context of curves only.

This change is required so border curve parameter can be interpreted on
surfaces and vis-versa.

This was clearly the intent as m_Surface.get_u0() was used to set_t0().