Update RefineISegChainSeg to handle border points differently.
ISeg points that are on a patch border do not do well with the
conventional surface-surface routine.

So we detect these cases, extract the border curve and then use a
curve-surface routine instead.

In addition, we check the before/after results and in all cases only
accept the result if it improves the distance between the points.