Make XSec Cut behave as expected with Linking and Advanced Linking
Before, cut actually cleared the 'clipboard' and moved the cross section
to the clipboard.  Therefore, the cross section being cut was not
immediately destroyed and its Parms were not destroyed, so the Links
were not invalidated.

This changes Cut to be implemented as a Copy followed by a Delete.  This
will always force a true delete, which will invalidate links.