Make ParasiteDrag t/c, finess ratio and FF calculations work with props
Propellers throw off the surface counting, this makes them match the
logic used by Lref to avoid wrong t/c and ff for components after props.

Propellers should generally be left out of the Set used for ParasiteDrag,
but this should make it safer.

It would be better to clean all this code up to avoid use of push_back
and to consolidate all the calculation stuff into a common routine.

Thanks to Jeff V. for pointing out this error.