Cache some calculations in hyperbolic tangent stretching function.
The Newton's method to solve asinc or asinhc is being needlessly repeated
for calls with repeated values of ds0, ds1.  This routine is frequently
called repeatedly in this manner.

A first attempt was made to produce a vector t version of this routine.
Unfortunately, the various call sites were too complex to make this
practical.

Instead, this caching approach limits changes to a single file and all
callers benefit.  In fact, there are some call sites with repeated calls
that are entirely un-suitable for vector calling that benefit from this.

This speedup is likely not significant (this has not been identified
as a problem through profiling), but the current code feels wasteful and
this fixes it.