Set up python virtual environment and test in CMake / CTest
This does not work with the address sanitizer in the build.

The dependencies are not set up, so equivalents to the following
steps must be run:

make
make venv
make test

I am not sure if changing various dependencies will trigger
updates of the virtual environment.  Ideally, changes to the
C++ source (that change the _vsp.so) and changes to the Python
source of the package *.py files would trigger updates.

In addition, the vspaero executables are not currently included
in the Python package.