OpenVSP 3.27.1
A new winner for the smallest OpenVSP release ever.  A single line changed.

As it turns out, the skinning GUI does not work in 3.27.0.  Some may find
this to be an improvement, but unfortunately it does count as a regression
and it was unintentional.

With much embarrassment, we will push out this .1 and hopefully everyone
can pretend like 3.27.0 never happened.  I hear people don't like .0
versions anyway.

Features
 - None

Fixes
 - Make skinning GUI work again