Handle property compatibility differently.
Using Choice::AddItem( string, int val ) with the val argument allows us
to work with Choice that does is not zero-indexed.

In the case of Fea Property choices, this allows us to build two Choice
pull-downs with either shell or cap properties in each list.

This is preferred to adding all the items and then marking some as
FL_MENU_INACTIVE as needed.  All the unavailable entries are ugly and
confusing.

In situations where the available options can come and go (like some
of the wing blending options), then a Choice item that turns on/off
is best.  However, if an option is never available, it should never be
seen.

This code is easier to read and maintain -- and the patch is a reduction
in code.