Write orthotropic materials to FEA files
In Calculix, this uses TYPE=ISO or TYPE=ENGINEERING CONSTANTS as
appropriate.

In NASTRAN, it uses MAT1 or MAT8.

MAT8 can only be used for shell elements.  So we either need to prevent
users from using orthotropic materials for beam elements, or provide
a way to write an approximately equivalent MAT1 for beam elements.