Only allow beam elements to choose isotropic materials    WARNING
This change restricts the GUI Choice for beam element materials to only
display isotropic materials.

This is reasonable for NASTRAN, but is inappropriate for CalculiX.

Instead of this restriction, we should find a way for orthotropic materials
to work with beam elements for NASTRAN.  This likely involves writing
an equivalent isotropic material property.