Improve FeaPart::Update()
Update() was a single call that had several pre- and post- steps.

Instead, we refactor in the pattern of Geom::Update() where the FeaPart
sublcasses must implement an UpdateSurface() method that pluggs into
a larger Update() framework implemented in FeaPart::Update()

This also starts the process of cleaning up how FEA related Update()
is called.