Split FeaMeshMgr::AddStructureParts() to cull parts for half mesh case
CleanMargeSurfs() culls any -Y surfaces when HalfMesh is true.

AddStructureParts() adds FeaParts both FixPoint and Surface types.

According to the comment, CleanMergeSurfs() needed to be before
AddStructureParts() for FEAFixPoint surface identification.

However, this meant that surface-based FeaParts were never culled if they
only existed in the -Y plane.

Here we split AddStructureParts() into the part that needs to be before
CleanMergeSurfs() (i.e. AddStructureSurfParts()) and the part that needs
to be after CleanMergeSurfs() (i.e. AddStructureFixPoints).

Both AddStructureXXX() are then simplified to remove un-used stuff that
was particular to the other part of the routine.