Speed up MergeIPntGroups, part of MergeBorderEndPoints
This performs a brute force O(N^2) search to find the nearest groups of
end points within a certain tolerance.

This change reduces the cost to O((N^2)/2) by recognizing that the
distance is symmetric and calculating both i,j and j,i is redundant.

A further improvement could be had by using nanoflann to perform an
accelerated nearest neighbor search.