Restore FEA_SKIN=7 and make new FeaPart types possible.
Previously, the code was fragile, requiring FeaParts to be added
in sequence to m_FeaPartChoice.  FEA_SKIN was not added, and SubSurface
parts followed.

Now, values are assigned and indexing is not used, this makes the code
robust to adding future part types -- so long as the current part type
numbers are not altered.

As a bonus, the code is considerably simplified in some places -- more
simplification is possible.