Remove un-used IntersectLineSegMesh
It looks like IntersectLineSegMesh was an alternate implementation of
IntersectLineSeg (which works on the Bezier surfaces).  The Mesh version
was included in CFDMesh as first introduced to v3 (history not traced
more).

As the Mesh version checks every triangle by brute force, it is likely
very slow and was probably never used.