Make Mesh::AddFace cause edges to track their faces
This requires adding Edge::RemoveFace.

This commit hides an additional impact.  Mesh::ConvertToQuads was not
tracking edges properly.  It either needed code like that removed here
or calls to Edge::SetFace().  Instead, this is handled by the change
to Mesh::AddFace (for quads).

So although this appears as a code simplification, it actually fixes a bug
where existing complexity was not properly implemented.