Protect for idouble > imax
On x86-64, the double to int conversion applied to the result of the
floor and ceil calls would lead to ifloor<0, and overall the code would
work.

However, on aarch64, the double to int conversion is handled differently,
leading to exceptionally large ifloor and iceil values that would blow
beyond the limits of target_vec, u_vec, and dist_vec -- causing invalid
memory access and a crash.

This change causes both platforms to behave the same any time idouble
is huge.  This actually happens because dsdi is near zero and we have
a near divide by zero.