Experimental:  Turn VSPCheckSetup() into a do-nothing
InitGui() got renamed to Init() which got renamed to VSPCheckSetup().

Although this gets called from many places, it only really runs once.

It makes sure there is a valid Vehicle pointer returned by VehicleMgr.

Perhaps this should be eliminated.  Perhaps the 'once' guard should be
removed so it checks every time it is called.

This is an experimental change to see what happens if it is just made
to do nothing.

This is likely far too aggressive for release, but it should help learn
about whether it is important.