Don't copy vspaero solver out for external project build.
The VSPAERO EP build is used when a different compiler is used for
just the solver (from the rest of the OpenVSP build).

In this case, there is no vspaero CMake TARGET at this level -- so the
add_custom_command(TARGET does not work.

These commands exist to copy the latest vspaero build to other directories
for test during development.  They are not needed for an automated
release build (which is what uses the EP build).