Remove VSPAERO Single analysis, mark AnalysisMgr analysis as deprecated.
The single analysis would run sweeps of VSPAERO, but just one at a time.
This put OpenVSP in charge of the sweep instead of VSPAERO.

This increased the computational time (as things like geometry manipulation
were done for each case) with the advantage of allowing the results manager
GUI to parse results as they became available.

SetDefaults() is called on startup, so we can't issue an error then as
it will get annoying.