Make ProcessUtil work with empty string path
A command executed with an empty path string should not have a / or \\
prepended -- as the operating system will interpret this as an absolute
path starting at the root.  Instead, a path-free command should be
executed in the current working directory -- and will probably only
work if the executable is in the operating system path.

On Windows, the current working directory is implicitly in the path whle
on Unix-type operating systems it is not.