Add routines to find CalculiX
We first check the specific path.  This will either be the location
of the vsp executable, or the location of the vsp shared library
in the case of the Python API.  In either case, this is the location
where the VSPAERO related binaries are expected.

If ccx is not found there, then the operating system's path is checked
to see if CalculiX has been installed for general use.

If found, m_CalculiXFound is set to true.

If found in the specific location, then m_CalculiXPath is set to that
location.  If it is found in the operating system's path, then this
variable is left as the empty string.