Make APIHideScreens() skip VSP_MAIN_SCREEN
Show()'ing the main screen after hiding causes OpenVSP to crash.  This
appears to be because FLTK destroys the OpenGL context when it closes
a FL_GL_Window and OpenVSP's OpenGL subsystem is not robust enough
to recreate everything when it is re-opened.

The prior code _only_ closed the main screen.  This was a debug hack.
where the screen was being closed/opened from the normal GUI.  It does
not make sense for the intended API use of this routine.