Download Eigen directly from GitLab.
Copying the tarball into the repo makes it difficult to verify the provenance
of the code. By pointing to the offical GitLab and recording the SHA1
sum it make it easier to verify that the code that actually came from
the official source. This also reduces the size of the repository
considerably.

Updates Eigen from 3.4-rc1 to 3.4.0