Don't update un-used variables in UpdateDisplaySettings.
UDS() exists to update the display settings used by UpdateDrawObj() and
LoadDrawObj() as the settings are changed in a post-intersection and mesh
generation timeframe.

All settings (not just the display ones) are copied at the start of the
intersection or mesh generation process via CopyFrom().

So anything that occurs non-interactively in the mesh gen process does
not need to be updated in UDS().  Only interactive things need to be here.