Add FEA specific WriteSTL implementation.
The CFDMeshMgr WriteSTL uses data stored in the surf->mesh, which does not
exist after a CAD-only run.

However, that data is not in a FeaMesh that will be stored when switching
between FeaStructures.

Consequently, writing STL files did not make sense associated with either
a CAD-only run, or a FeaMesh.  Adding a FeaMesh::WriteSTL solves this
issue.