Improve border fixed point handling.
Previously, only borders where UW for each patch was equal would be found.

This missed W=0,W=1 border as well as seams at U=0 or U=1 (along wingtips).

This change is designed to leave the old logic as-is as much as possible,
but if a fixed point is found to be on a border -- but no matching patch
is identified, we try harder to find the matching patch.