Fix problem where intersection fixed points were getting ignored.
This was broken in bb6e72e76e, where fixed point storage was changed
from a bunch of arrays to a single array of objects.  Modifying the
array updated the authoritative version of the fixed point, whereas
modifying fxpt only updates a copy.

The other appearances of fxpt on the LHS of an assignment are not needed.