Make sure forked process runs for at least one second to allow monitor.
When we fork a process -- like VSPAERO or CalculiX, we also fork a
monitor process to watch the first process' stdout and redirect it
to the console.

However, if the process terminates before the monitor is operating, then
we capture no output.

This would occur when CalculiX would crash -- giving no messages to help
debug the problem.

This change makes sure we capture any messages from a failing program.