Force NACA 6A Series airfoils to use a=0.8 camber line.
NACA 903 makes clear that only the a=0.8 camber line should be used with
the 6A series.  The 6A series was designed to provide straight (rather
concave cusped) surfaces near the trailing edge.  The 6A foils with a=0.8
use a modified camber line that does not ruin the straight surfaces.

Using any other camber line will not achieve the straight surfaces that are
the point of a 6A series foil.

Thanks Daniel Mesward