OpenVSP 3.31.0
OpenVSP Halloween Edition!

For those of you keeping track, this ominous release was foreshadowed
back in the 3.28.0 release notes.  Like that release, this one is all
about structures.  There are a few other things, but mostly structures.

That isn't to say if you don't use OpenVSP's structures capability you
shouldn't update -- you most certainly should.

This release includes one very subtle feature that should impact every
single user.  It is worth updating for this feature alone!

The catch?  I'm not going to tell you what it is.  We'll see who is
the first to figure it out.  Post your guesses on LinkedIn or the
Google Group.

First, some of the non-structures updates...

NACA 6A Series airfoils will now be restricted to the modified a=0.8
camber line.  Although in many cases we go beyond the strict NACA airfoil
definitions and allow things their creators did not intend, in this case
it seems appropriate to restrict ourselves.  If you consider this a bug
fix, it has been with us since the first 6-series support added in RAM
v2.1 in the early 2000's.

The Analysis menu has received some re-organization and renaming to
facilitate all the structures work.  There are now "Structure" and
"Aero" sub-menus.  Hopefully everyone's muscle memory will adjust
without too much trouble.

The CMake build system has been extended to build RPM files for RedHat
type distributions.  We are not yet providing pre-built RPM's as our
CI build servers only run Ubuntu.  This should be useful for anyone
supporting multiple RPM based machines.

There are a bunch of VSPAERO fixes and cleanups.  Nothing really major
here, but a whole bunch of little things from the team.

And the structures updates...

The FEA Mesh workflow has been broken apart.  You can now generate
a mesh without writing it to file -- and write a previously generated
mesh to file at a later time.  This added flexibility was required for
some of the later improvements, but come as a nice addition on their own.

Structural boundary conditions can now be specified for each structure.
BC's are applied to mesh nodes.  The user selects which DOF to restrict
and also how to identify which nodes to apply each BC to.

Structural assemblies have been added.  Each assembly is made of one or
more structures.  These structures can be connected with rigid links
between Fixed Points in each structure.  Assemblies are written out to
a combined FEM file with node and element offsets automatically generated.

A VSPAERO / CalculiX based Aero-Structure workflow GUI has been added.
This GUI helps guide the user through a process of generating loads,
applying them to a structure, and then analyzing the result.

Note:  This is the final release where an Ubuntu 18.04 package will be
built for download.  Please update to 20.04 or 22.04.

Features:
 - Trick or treat?  Halloween mystery feature...
 - Add structural assemblies made from multiple structures.
 - Add rigid connections between structures in assembly.
 - Add boundary condition specification to structures.
 - Add aero-structure coupled problem GUI using CalculiX and VSPAERO.
 - Split FEA Mesh generation from writing to file.
 - Also split FEA CAD file generation and export from Meshing.
 - Keep FEA Mesh in-memory, allowing more flexible workflow.
 - Re-arranged menus to group structures and aero analyses.
 - Build RPM packages for RedHat-based distributions.  Thanks Jeff A.
 - Add aerodynamic center stability mode to GUI.
 - Force NACA 6A series to use a=0.8 camber line.  Thanks Daniel M.
 - Remove VSPAERO single analysis from API - reduce code duplication.

Bug Fixes:
 - Allow non-unique names in pull-down menus.  Thanks Brandon.
 - Make environment.yml match Python version.  Thanks Jason W.
 - Fix component IDs passed to *.vspgeom.  Fixes propellers.
 - Write NORMAL at all B32 nodes, not just the first.
 - Fix exporting symmetrical propeller at origin to STL.  Thanks Brandon.
 - Insert FixPoins before InitMesh -- guarantee Fix Points
 - Lots of VSPAERO fixes
 - Lots of structures fixes