Re-tag CompGeom mesh after half-mesh to eliminate empty tags.
Empty tags in the *.vkey file could cause the indexes in that file to
be non-consecutive (in the non-empty cases).  This would cause VSPAERO
to crash as it would index past the end of an ID array.

Re-tagging the mesh re-builds the tag map held by SubSurfMgr and
eliminates any empty tags.

In order for re-tagging to work, several of the arrays and maps had to
be cleared before re-use.