Call RemapID on all remaining ID's in DecodeXml.
ID remapping is used to resolve ID collisions when inserting a file or
copy/pasting entities.  The new entity (with colliding ID's) has its IDs
remapped on the fly to new ones.

Most of these will not make any difference because they are not entities
that are inserted or copy/pasted.  However, it is good practice to
call RemapID on all ID's just to be sure.

VarPresetMgr does participate in insertion, so the remap on p_IDs may
actually be a bug fix.