Use non-flattened static call to ErrorMgr getInstance()
SWIG 4.1.0 Issue 2137 eliminates the creation of 'flattened' wrappers
for static methods by default for Python.

This can be changed with the -flatstaticmethod flag -- but it was not
obvious how to apply that flag to the right context in CMake.

So, instead, the recommended procedure is to use the . instead of _ to
call getInstance().