Structures materials and properties update.
Track material and property association by ID instead of index.

Handle units of materials and properties very differently.

Update built-in material values (metals).


Consolidate FeaPart to use KeepDel and CreateBeam flags instead of
IncludeElements.  This parallels the changes that were required for
subsurfaces (but optional here).  It simplifies the GUI in the process.