OpenVSP 3.32.2
While 3.32.1 fixed the most egregious issues with 3.32.0, it was still
a bit of a hot mess in its own right.  This release helps tame the chaos,
hopefully it gets people back on track.

Of course, I couldn't help myself and stick to _just_ bug fixes.  So,
if you use CFDMesh & Co, you will find that the 'f' and 'r' keys work
like they do in the rest of OpenVSP.  These are UX candy -- go ahead
and indulge.

(For those of you that are wondering what pressing 'f' or 'r' does --
try it!  If you can't try it right now, pressing 'f' will 'f'it all the
visible stuff onto the screen by auto zoom, pan, and setting the center
of rotation.  Pressing 'r' presents a bunch of green points.  Click one
and it will be set to the center of rotation and moved to the center of
the screen.

Other than that, it should mostly be a bunch of fixes.  The migration
from sprintf to snprintf broke a few things.  Those have been fixed and
an audit has been conducted, so hopefully we're through with that.

Fix lingering problems with CompGeom half meshes and tag files.  Previous
fixes broke things like Wave Drag and Parasite Drag.

Fix error validating subsurface structures properties on structures
data transfer.  Was preventing analysis of structures with subsurfaces.
Improve error messages while we're at it.

Attempt to set property length unit when converting an old file.
Unfortunately, I can't see a way to convert

Some of the VSPAERO output files were updated, our parsing of those files
has now been updated to match.

Fix some small VSPAERO issues that were causing hangs and crashes when
slicing planes were active.

Overall, not a ton of stuff -- but everyone should update because the
problems these fixes address were hitting a bunch of users.  Hopefully
we'll keep things under control for a while going forward.

Features
 - Make 'f' and 'r' keys work with CFDMesh & Co.
 - Allow parasite drag to use a reference wing not in the active Set.
 - Print missing command line when calling CpSlicer.
 - Clean up mesh from repeated mass properties analysis.
 - Attempt to set property length on import of older structures files.

Fixes
 - Fix validating structures subsurfaces
 - Fix visual artifacts in orthotropic material GUI in structures.
 - Fix double fee in Cartesian plotting library that shows up with bleeding
   edge versions of FLTK.
 - Fix parsing of VSPAERO output files.
 - Fix wave drag, parasite drag, half mesh, tag files issue.
 - Fix snprintf conversion issues
 - Silence snprintf warnings
 - Fix VSPAERO with slicing planes active